# Fixes issue #6071 - TIFF with 1 bit-depth