--- a/test/pdfs/issue6071.pdf.link
+++ b/test/pdfs/issue6071.pdf.link
@@ -1,0 +1,1 @@
+https://web.archive.org/web/20170107214304/https://www.pdf-archive.com/2017/01/07/issue6071/issue6071.pdf