@jazzchipc 2016-12-06
Hello timvandermeij and Snuffleupagus. Thanks for the feedback. I rearranged some things and forgot about the code style. I've corrected both the identation and the commit message as you asked.

Are there any more things you want to point out or know about the code?