@Snuffleupagus 2016-12-06
> Are there any more things you want to point out or know about the code?

I'd suggest that you remove the `Corrected sintaxe according to style guide.` part of the commit message, since that's unnecessary (and also, please note that the correct spelling is `syntax` :-)

Unfortunately I don't know the TIFF code at all, so I'm not sure that I'd be able to do a meaningful code review here. Perhaps @yurydelendik can help with that, if he has time!?