@jazzchipc 2016-12-17
Hi again, sorry for the delay in answering. 

@yurydelendik I took your approach and created a slightly more efficient version. When you told me to eliminate the inner cycle I didn't really know if you meant what I wrote or creating a pre-mapping array.

In this next table, baseline is your code and current is my commit:

| browser | stat | Count | Baseline(ms) | Current (ms) | +/- | % |
| ---- | ---- | ---- | ---- | ---- | ---- | ---- |
| Firefox | Overall | 20 | 113| 103| -10| -9.09 |
| Firefox | Page Request | 20 | 12 | 9 | -3 | -22.50|
| Firefox | Rendering | 20 | 101 | 93 | -7| -7.46 |

I don't know if my testing conditions were different than the initial commit, because these values are higher than the ones I first posted... But testing with my first commit, this code was 14% faster.

Also, about the PhotometricInterpreation thing, I'll try to test it, but I'm having problems:

I can create a TIFF image with 1-bit depth and 1 color, but I can't convert it to .pdf format and "activate" this block we're testing. Any chances you know how to solve this and convert/incorporate the TIFF image correctly?