@yurydelendik 2016-12-19
Just some proof:
![gif.latex](https://cloud.githubusercontent.com/assets/1523410/21316473/895e6282-c5c6-11e6-825d-d23ea432673e.gif)