@yurydelendik 2016-12-19
> Also, about the PhotometricInterpreation thing, I'll try to test it,

I would like to keep PhotometricInterpreation out-of-scope of this PR.