@jazzchipc 2016-12-21
@yurydelendik Sorry for not explaining the optimization. I couldn't figure it out alone either, but got help from Stack Overflow because I believed there was a way to do it. Turns out it was quite simple. Not a huge jump in optimization, but still better.

Let me know if there's anything else to change. I'll open a PR for PhotometricInterpretation once I can test it.