@yurydelendik 2016-12-21
/botio test

can you add a reference test?