@jazzchipc 2016-12-21
@yurydelendik How could I add a test? And should I use the pdf of the issue? Do we have the copyrights to use it?