@yurydelendik 2016-12-21
You will need to modify test_manifest.json file (see it's history https://github.com/mozilla/pdf.js/commits/master/test/test_manifest.json for examples). Depending on if pdf's origin is unknown (or copyrighted) and we does not know if it's allowed to distribute it, we just including the reference (via webarchive?) -- link test. But we prefer if pdf author agree to include the example and we can add it to the repo.