@timvandermeij 2016-12-21
Example: https://github.com/mozilla/pdf.js/commit/c5b06cb40d032e782f5afa3a0ed73d4ee80f160a (only the changes for files in the `test/` folder.