@jazzchipc 2016-12-22
Should I make another commit or squash the addition of the test? Btw, should I upload the PDF itself or a webarchive link to it? If I point the link, can it still be added to test_manifest?