@timvandermeij 2016-12-23
You should squash the test into the existing commit. Since we do not have explicit permission to add the PDF to the repository, a Web Archive link would be good. Example of a test manifest entry with a Web Archive link: https://github.com/mozilla/pdf.js/commit/54ee83eb125ff6bbf95a4e7a0e7ef2493e4aae4d