@jazzchipc 2016-12-24
I tried doing that, but since the PDF is hosted on DropBox, I cannot save a link to it on WebArchive, unless I reupload it somewhere else.