@jazzchipc 2017-01-10
I believe I added the test correctly. Please review the new commit and let me know if something is wrong or should be added.