@Snuffleupagus 2017-01-10
Note that the Linux bot failed to download the new file (but it worked on the Windows bot):
```
Downloading https://web.archive.org/web/20170107214304/https://www.pdf-archive.com/2017/01/07/issue6071/issue6071.pdf to pdfs/issue6071.pdf...
Error during downloading of https://web.archive.org/web/20170107214304/https://www.pdf-archive.com/2017/01/07/issue6071/issue6071.pdf: HTTP 503
WARNING: File was not downloaded. See "pdfs/issue6071.pdf.error" file.
```