@jazzchipc 2017-01-10
Says it was the checksum that failed, but why would it fail on Linux and now on Windows?