@jazzchipc 2017-01-10
Linux bot says the checksum is different, but I checked on Linux just now, with the md5sum command, and it gives me the same checksum as Windows. But I'm checking other .pdf.link files and they all have the first line with the link and a line break, making it 2 lines. Mine only has 1. Should I fix that and retry?