@yurydelendik 2017-01-10
Okay, I placed it in the linux's pdf-cache