@jazzchipc 2017-01-10
Thanks for guiding me in here. I'm glad I could help even though it was a small contribution. ;)