@bfarmilo 2017-01-31
Hi - I seem to be getting the same inversion issue when using the svg renderer -- the test file (and all of my other images) come out with white rendered as black and black rendered as white. Before I post additional details, should this fix described above also work for SVG or is the TIFF handling code completely separate ? So far in all of my digging this thread has come closest to describing the problem. Note that when I use the canvas approach it comes out perfectly.

Happy to post my code and examples, please let me know if this is an appropriate place for this !

Thanks