@timvandermeij 2017-01-31
It looks like this is unrelated to this PR. Please open a new issue with the requested information.